Time: 0 ms (100%), Space: 8.7 MB (35.91%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if(x<0){return false;}
-        int revNum = 0;
+        long revNum = 0;
         int copy = x;
         while(x != 0){
             int last_digit = x%10;
